# frontend/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AwardCount": {
        "properties": {
          "award_type_code": {
            "title": "Award Type Code",
            "type": "string"
          },
          "award_type_id": {
            "title": "Award Type Id",
            "type": "integer"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "award_type_id",
          "award_type_code",
          "count"
        ],
        "title": "AwardCount",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlayerCreate": {
        "additionalProperties": false,
        "properties": {
          "cohort_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cohort Id"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "first_name": {
            "maxLength": 50,
            "minLength": 1,
            "title": "First Name",
            "type": "string"
          },
          "joined_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Joined Date"
          },
          "last_name": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "squad_number": {
            "anyOf": [
              {
                "maximum": 99.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Squad Number"
          },
          "team_season_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Season Id"
          }
        },
        "required": [
          "first_name"
        ],
        "title": "PlayerCreate",
        "type": "object"
      },
      "PlayerMove": {
        "additionalProperties": false,
        "description": "Move a player between two teams' squads (same cohort) - an age-group coach action.",
        "properties": {
          "from_team_season_id": {
            "title": "From Team Season Id",
            "type": "integer"
          },
          "left_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Left At"
          },
          "squad_number": {
            "anyOf": [
              {
                "maximum": 99.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Squad Number"
          },
          "to_team_season_id": {
            "title": "To Team Season Id",
            "type": "integer"
          }
        },
        "required": [
          "from_team_season_id",
          "to_team_season_id"
        ],
        "title": "PlayerMove",
        "type": "object"
      },
      "PlayerRead": {
        "properties": {
          "cohort_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cohort Id"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "joined_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Joined Date"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "left_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Left Date"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "required": [
          "id",
          "cohort_id",
          "first_name",
          "last_name",
          "display_name",
          "date_of_birth",
          "joined_date",
          "left_date",
          "notes"
        ],
        "title": "PlayerRead",
        "type": "object"
      },
      "PlayerStatsRow": {
        "properties": {
          "appearances": {
            "title": "Appearances",
            "type": "integer"
          },
          "assists": {
            "title": "Assists",
            "type": "integer"
          },
          "awards": {
            "items": {
              "$ref": "#/components/schemas/AwardCount"
            },
            "title": "Awards",
            "type": "array"
          },
          "goals": {
            "title": "Goals",
            "type": "integer"
          },
          "goals_per_game": {
            "title": "Goals Per Game",
            "type": "number"
          },
          "minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minutes"
          },
          "own_goals": {
            "title": "Own Goals",
            "type": "integer"
          },
          "player": {
            "$ref": "#/components/schemas/PlayerSummary"
          },
          "squad_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Squad Number"
          },
          "starts": {
            "title": "Starts",
            "type": "integer"
          }
        },
        "required": [
          "player",
          "squad_number",
          "appearances",
          "starts",
          "goals",
          "assists",
          "own_goals",
          "goals_per_game",
          "awards",
          "minutes"
        ],
        "title": "PlayerStatsRow",
        "type": "object"
      },
      "PlayerSummary": {
        "description": "Lightweight embed for nested responses.",
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "display_name"
        ],
        "title": "PlayerSummary",
        "type": "object"
      },
      "PlayerUpdate": {
        "additionalProperties": false,
        "properties": {
          "date_of_birth": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "first_name": {
            "anyOf": [
              {
                "maxLength": 50,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "joined_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Joined Date"
          },
          "last_name": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "left_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Left Date"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "title": "PlayerUpdate",
        "type": "object"
      },
      "PositionCategory": {
        "enum": [
          "GK",
          "DEF",
          "MID",
          "FWD"
        ],
        "title": "PositionCategory",
        "type": "string"
      },
      "PositionRead": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/PositionCategory"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sort_order": {
            "title": "Sort Order",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "category",
          "sort_order"
        ],
        "title": "PositionRead",
        "type": "object"
      },
...
      "SquadMemberRead": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "joined_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Joined At"
          },
          "left_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Left At"
          },
          "player": {
            "$ref": "#/components/schemas/PlayerRead"
          },
          "primary_position": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PositionRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "squad_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Squad Number"
          },
          "team_season_id": {
            "title": "Team Season Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "team_season_id",
          "player",
          "squad_number",
          "primary_position",
          "joined_at",
          "left_at"
        ],
        "title": "SquadMemberRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ABGFC Blues",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/players": {
      "get": {
        "operationId": "list_players_api_v1_players_get",
        "parameters": [
          {
            "in": "query",
            "name": "cohort_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cohort Id"
            }
          },
          {
            "in": "query",
            "name": "include_left",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Left",
              "type": "boolean"
            }
          },
          {
            "in": "cookie",
            "name": "abgfc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Abgfc Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlayerRead"
                  },
                  "title": "Response List Players Api V1 Players Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Players",
        "tags": [
          "players"
        ]
      },
      "post": {
        "operationId": "create_player_api_v1_players_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "abgfc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Abgfc Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Player",
        "tags": [
          "players"
        ]
      }
    },
    "/api/v1/players/{player_id}": {
      "get": {
        "operationId": "get_player_api_v1_players__player_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "title": "Player Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "abgfc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Abgfc Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Player",
        "tags": [
          "players"
        ]
      },
      "patch": {
        "operationId": "update_player_api_v1_players__player_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "title": "Player Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "abgfc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Abgfc Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Player",
        "tags": [
          "players"
        ]
      }
    },
    "/api/v1/players/{player_id}/memberships": {
      "get": {
        "description": "Every team-season the player has been in (that the caller can see).",
        "operationId": "player_memberships_api_v1_players__player_id__memberships_get",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "title": "Player Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "abgfc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Abgfc Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SquadMemberRead"
                  },
                  "title": "Response Player Memberships Api V1 Players  Player Id  Memberships Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Player Memberships",
        "tags": [
          "players"
        ]
      }
    },
    "/api/v1/players/{player_id}/move": {
      "post": {
        "operationId": "move_player_api_v1_players__player_id__move_post",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "title": "Player Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "abgfc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Abgfc Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerMove"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SquadMemberRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Move Player",
        "tags": [
          "players"
        ]
      }
    },
    "/api/v1/players/{player_id}/stats": {
      "get": {
        "operationId": "player_stats_api_v1_players__player_id__stats_get",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "title": "Player Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "team_season_id",
            "required": true,
            "schema": {
              "title": "Team Season Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "abgfc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Abgfc Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerStatsRow"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Player Stats",
        "tags": [
          "players"
        ]
      }
    },
...
  }
}